Windows build script. Specific known working versions of SDL2 and libtcod checked in with build scripts. Compiled DLLs, static libraries and modaccent executables are all checked in with PDBs.
Various fixes for things g++ errors on that cl.exe does not

Ignore `Win32_<Debug|Release>` directory created by `run.bat`.

As esran points out in the d3d460ebc7ae86af9151ece7d6433433623c0dca we should really be recompiling the Incursion code after using it to compile the module, so that `VMachine.cpp` gets a correct `dispatch.h` for the current module code. Note to the reader, we need to note this in the readme.

Api.h cannot use `size_t` because our script language does not support it. The intention was to fix the build warnings, but it will cause module compilation to error on the line with the unrecognised `size_t` typed parameter to `random()`.

+++ inc/Inline.h
@@ -38,7 +38,7 @@
 inline int16 random(int16 mx)
   {
     if (!mx) return 0;
-    return (genrand_int32() % mx);
+    return (int16)(genrand_int32() % mx);
   }
   
 #define srand(n) init_genrand(n)
